pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "dateparser>=1.2.1",
    "discord-py>=2.5.2",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
]
